video user and tweet controllers complete

+++ src/controllers/tweet.controller.js
@@ -1,3 +1,4 @@
+import mongoose, { isValidObjectId } from "mongoose";
 import { Tweet } from "../models/tweet.model.js";
 import { ApiError } from "../utils/ApiError.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
@@ -27,29 +28,112 @@ const createTweet = asyncHandler(async (req, res) => {
     .status(200)
     .json(new ApiResponse(200, tweet, "Tweet created successfully"));
 });
-//Testing
+
 const getUserTweets = asyncHandler(async (req, res) => {
-  const userTweets = await Tweet.aggregate([
+  const { userId } = req?.params;
+
+  if (!isValidObjectId(userId)) {
+    throw new ApiError(400, "Invalid user Id");
+  }
+
+  const tweets = await Tweet.aggregate([
     {
       $match: {
-        owner: req?.user?._id,
+        owner: new mongoose.Types.ObjectId(userId),
+      },
+    },
+    {
+      $lookup: {
+        from: "users",
+        localField: "owner",
+        foreignField: "_id",
+        as: "ownerDetails",
+        pipeline: [
+          {
+            $project: {
+              avatar: 1,
+              userName: 1,
+              fullName: 1,
+            },
+          },
+        ],
+      },
+    },
+    {
+      $lookup: {
+        from: "likes",
+        localField: "_id",
+        foreignField: "tweet",
+        as: "likeDetails",
+        pipeline: [
+          {
+            $project: {
+              likedBy: 1,
+            },
+          },
+        ],
+      },
+    },
+    {
+      $addFields: {
+        likesCount: {
+          $size: "$likeDetails",
+        },
+        ownerDetails: {
+          $first: "$ownerDetails",
+        },
+        isLiked: {
+          $cond: {
+            if: { $in: [req?.user?._id, "$likeDetails.likedBy"] },
+            then: true,
+            else: false,
+          },
+        },
+      },
+    },
+    {
+      $sort: {
+        createdAt: -1,
+      },
+    },
+    {
+      $project: {
+        content: 1,
+        ownerDetails: 1,
+        likesCount: 1,
+        isLiked: 1,
+        createdAt: 1,
       },
     },
   ]);
 
-  if (!userTweets) {
-    throw new ApiResponse(300, {}, "No tweet found");
+  if (!tweets || tweets.length === 0) {
+    throw new ApiError(300, "This user has no tweet");
   }
+
   return res
     .status(200)
-    .json(new ApiResponse(200, userTweets, "Fetched user tweets successfully"));
+    .json(new ApiResponse(200, tweets, "Fetched user tweets successfully"));
 });
-//Testing
+
 const updateTweet = asyncHandler(async (req, res) => {
   const { tweetId } = req?.params;
   const { content } = req?.body;
   if (!tweetId) {
     throw new ApiError(400, "Tweet id not found");
+  }
+
+  const tweet = await Tweet.findById(tweetId);
+  if (!tweet) {
+    throw new ApiError(400, "No tweet found with this id");
+  }
+  console.log(new mongoose.Types.ObjectId(tweet?.owner));
+  console.log(req?.user?._id);
+  if (req?.user?._id?.toString() !== tweet?.owner?.toString()) {
+    throw new ApiError(
+      400,
+      "Only user who created this tweet can update this tweet",
+    );
   }
 
   const updatedTweet = await Tweet.findByIdAndUpdate(
@@ -65,15 +149,27 @@ const updateTweet = asyncHandler(async (req, res) => {
   if (!updatedTweet) {
     throw new ApiError(500, "Error while updating tweets");
   }
+
   return res
     .status(200)
     .json(new ApiResponse(200, updatedTweet, "Tweet updated successfully"));
 });
-//Testing
+
 const deleteTweet = asyncHandler(async (req, res) => {
   const { tweetId } = req?.params;
   if (!tweetId) {
     throw new ApiError(400, "Tweet id not found");
+  }
+  const tweet = await Tweet.findById(tweetId);
+  if (!tweet) {
+    throw new ApiError(400, "Cannot find tweet with this Id");
+  }
+
+  if (tweet?.owner?.toString() !== req?.user?._id?.toString()) {
+    throw new ApiError(
+      400,
+      "Only user who created this tweet can delete this tweet",
+    );
   }
   const tweetToBeDeleted = await Tweet.findByIdAndDelete(tweetId);
   if (!tweetToBeDeleted) {
@@ -84,4 +180,4 @@ const deleteTweet = asyncHandler(async (req, res) => {
     .json(new ApiResponse(200, {}, "Tweet deleted successfully"));
 });
 
-export { createTweet, getUserTweets, updateTweet , deleteTweet };
+export { createTweet, getUserTweets, updateTweet, deleteTweet };
